backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
motor==3.7.0
pydantic-settings==2.9.1
python-jose[cryptography]==3.3.0
APScheduler==3.10.4
httpx==0.28.1
python-dotenv==1.1.0
faker==37.1.0

